pyproject.toml
[tool.poetry]
name = "language-model-repair"
version = "0.1.0"
description = ""
authors = ["tvergara <tomvergara@uc.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter-client = "^8.6.3"
pynvim = "^0.5.0"
jupyter = "^1.1.1"
torch = "^2.4.1"
transformers = "^4.45.2"
tqdm = "^4.66.5"
einops = "0.4.0"
higher = "0.2.1"
hydra-core = "1.2.0"
matplotlib = "3.6.1"
spacy = "3.4.1"
scipy = "1.9.2"
scikit-learn = "1.0.2"
nltk = "3.7"
datasets = "^3.0.1"
evaluate = "^0.4.3"
bert-score = "^0.3.13"
python-dotenv = "^1.0.1"
peft = "^0.14.0"
lightning = "^2.4.0"
comet-ml = "3.35"
pynvml = "^12.0.0"
pytest = "^8.3.4"
pylint = "^3.3.3"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
